FIX: lazy Suspense를 위해 todoDetail 컴포넌트로 이동
